Added FontAwesome icons
It's a pity about the extra overhead but the fontawesome icon set is just so much more complete than the Boostrap included glyphicons and look noticaebly nicer.

Also the 2 Javascript dependencies could go but then the dropdown menu in mobile view mode would not work. Anyway, they are from a CDN so in theory they should only be loaded once and sometimes not even that once if someone else has already used the same CDN assets on some other site.