Slightly improved the frontpage link columns width
Thanks to Mr Arisonic for suggesting to widen the navigation link columns. I was going to make it just a single column to accommodate some of the extra wide links and noticed that the column width class was set to col-lg-4 instead of col-lg-6 so we will see how this goes and maybe remove some of the longer links anyway.